Refactor: split large files, add shared hex_utils
