+++ main/hex_utils.c
@@ -1,0 +1,29 @@
+#include "hex_utils.h"
+#include <string.h>
+#include <stdio.h>
+
+int hex_digit(char c) {
+    if (c >= '0' && c <= '9') return c - '0';
+    if (c >= 'a' && c <= 'f') return c - 'a' + 10;
+    if (c >= 'A' && c <= 'F') return c - 'A' + 10;
+    return -1;
+}
+
+int hex_to_bytes(const char *hex, uint8_t *out, size_t out_len) {
+    size_t hex_len = strlen(hex);
+    if (hex_len % 2 != 0 || hex_len / 2 > out_len) return -1;
+    for (size_t i = 0; i < hex_len / 2; i++) {
+        int hi = hex_digit(hex[2 * i]);
+        int lo = hex_digit(hex[2 * i + 1]);
+        if (hi < 0 || lo < 0) return -1;
+        out[i] = (uint8_t)((hi << 4) | lo);
+    }
+    return (int)(hex_len / 2);
+}
+
+void bytes_to_hex(const uint8_t *bytes, size_t len, char *out) {
+    for (size_t i = 0; i < len; i++) {
+        sprintf(out + 2 * i, "%02x", bytes[i]);
+    }
+    out[len * 2] = '\0';
+}
